a lot of changes, pending refactor all dtos
