+++ core/domain/pawnOrder.go
@@ -1,20 +1,18 @@
 package domain
 
-import (
-	"time"
-)
+import "time"
 
 type PawnOrder struct {
-	ID         uint      `json:"id"`
-	EmployeeID uint      `json:"employee_id"`
-	Employee   *Employee `json:"employee"`
-	CustomerID uint      `json:"customer_id"`
-	Customer   *Customer `json:"customer"`
-	StatusID   uint      `json:"status_id"`
-
-	TotalMount    float64   `json:"total_mount"`
+	ID            uint      `json:"id"`
+	EmployeeID    uint      `json:"employee_id"`
+	Employee      *Employee `json:"employee,omitempty"`
+	CustomerID    uint      `json:"customer_id"`
+	Customer      *Customer `json:"customer,omitempty"`
+	StatusID      uint      `json:"status_id"`
+	Status        *Status   `json:"status,omitempty"`
+	TotalAmount   float64   `json:"total_amount"`
 	Monthly       bool      `json:"monthly"`
 	Products      []Product `json:"products"`
-	CutOffDay     time.Time
+	CutOffDay     time.Time `json:"cutoff_date"`
 	ExtensionDate time.Time `json:"extension_date"`
 }
