+++ core/infrastructure/persistance/employeePostgres.go
@@ -62,7 +62,7 @@ func (r *EmployeeRepositoryImpl) GetByEmail(c *gin.Context) (*domain.Employee, e
 func (r *EmployeeRepositoryImpl) GetAll() (*[]domain.Employee, error) {
 	var employees []domain.Employee
 
-	if err := r.database.Model(domain.Employee{}).Find(&employees).Error; err != nil {
+	if err := r.database.Model(domain.Employee{}).Preload("Profile").Find(&employees).Error; err != nil {
 		return nil, errors.New("failed to establish a connection with our database services")
 	}
 
@@ -78,7 +78,7 @@ func (r *EmployeeRepositoryImpl) Create(c *gin.Context) (*domain.Employee, error
 	auth.EncryptPassword([]byte(employee.Password), crypt)
 	employee.Password = string(<-crypt)
 
-	if err := r.database.Model(&domain.Employee{}).Create(&employee).Error; err != nil {
+	if err := r.database.Model(&domain.Employee{}).Omit("PawnOrdersDone").Omit("SessionLogs").Omit("EndorsementsDone").Create(&employee).Error; err != nil {
 		return nil, errors.New("failed to establish a connection with our database services")
 	}
