+++ core/interfaces/employeeHandler.go
@@ -4,7 +4,6 @@ import (
 	"net/http"
 
 	"github.com/JairDavid/Probien-Backend/core/application"
-	"github.com/JairDavid/Probien-Backend/core/infrastructure/auth"
 	"github.com/JairDavid/Probien-Backend/core/interfaces/common"
 	"github.com/gin-gonic/gin"
 )
@@ -13,7 +12,7 @@ func EmployeeHandler(v1 *gin.RouterGroup) {
 
 	interactor := application.EmployeeInteractor{}
 	employeeHandlerV1 := *v1.Group("/employees")
-	employeeHandlerV1.Use(auth.RoutesAndAuthority(true))
+	//employeeHandlerV1.Use(auth.RoutesAndAuthority(true))
 	employeeHandlerV1.POST("/", func(c *gin.Context) {
 		employee, err := interactor.Create(c)
