+++ router/setup.go
@@ -1,7 +1,6 @@
 package router
 
 import (
-	"github.com/JairDavid/Probien-Backend/core/infrastructure/auth"
 	"github.com/JairDavid/Probien-Backend/core/interfaces"
 	"github.com/gin-gonic/gin"
 )
@@ -10,7 +9,7 @@ func Setup(server *gin.Engine) {
 	v1 := *server.Group("/probien/api/v1")
 	interfaces.AuthHandler(&v1)
 
-	v1.Use(auth.AuthJWT())
+	//v1.Use(auth.AuthJWT())
 	{
 		interfaces.EmployeeHandler(&v1)
 		interfaces.BranchOfficeHandler(&v1)
